Option to start the REPL from the buffer directory
Defaults to on. The way it works is by changing Vim's current
working directory to match the directory of the file buffer
before starting the REPL. It changes the cwd back to what it
originally was to minimize side effects. This can make it easier
for the users code to use path relative to that script.