Option to start the REPL from the buffer directory (#32)
* Option to start the REPL from the buffer directory

Defaults to on. The way it works is by changing Vim's current
working directory to match the directory of the file buffer
before starting the REPL. It changes the cwd back to what it
originally was to minimize side effects. This can make it easier
for the users code to use path relative to that script.

* Documenation fix: codi#use_buffer_dir is a link

* White space / indentation fixes in documentation

* Improved code comments:

Improved grammar, capitalization, and spacing

* Better nvim implementation of codi#use_buffer_dir:

When using Neovim, this feature is implemented by passing a cwd
option to jobstart() instead of temporarilly changing Vim's current
working directory. This is less likely to cause problems.

* Improved doc for g:codi#user_buffer_dir():

Now it describes what instead of how, which cuases it to no longer
be inaccurate for Neovim.

* use_buffer_dir option can also be overridden on a per interpreter basis