restored normal node global context behavior:
Now that custom node REPL behavior is specified, it is a good idea
to enable the global context. If the standard REPL was being used,
the global context would be used. Also, when not running as a REPL,
node normally uses this context, so this option prevents the REPL
from further deviating from normal node behavior. Some JavaScript
code relies on the global context to work correctly.